tests functions extraction
