+++ tests/api/activitypub/actor.test.js
@@ -1,16 +1,11 @@
-const CONFIG = require('config')
-const _ = require('builders/utils')
 require('should')
-const { createUser, createUsername, createUserOnFediverse } = require('../fixtures/users')
+const { createUsername, createUserOnFediverse } = require('../fixtures/users')
 const { signedReq } = require('../utils/utils')
-const { startActivityPubServer } = require('../utils/activity_pub')
+const { query, startServerWithEmetterUser, createReceiver, makeActorUrl } = require('../utils/activity_pub')
 const { rawRequest } = require('../utils/request')
 const { shouldNotBeCalled, rethrowShouldNotBeCalledErrors } = require('../utils/utils')
-const { wait } = require('lib/promises')
 
-const host = CONFIG.fullHost()
 const endpoint = '/api/activitypub'
-const query = username => `${endpoint}?action=actor&name=${username}`
 
 describe('activitypub:actor', () => {
   it('should reject unsigned request', async () => {
@@ -35,8 +30,8 @@ describe('activitypub:actor', () => {
       const emetterUser = await createUserOnFediverse()
       delete emetterUser.publicKey
       const emetterActorUrl = await startServerWithEmetterUser(emetterUser)
-      const receiverActorUrl = await createReceiverActorUrl({ fediversable: false })
-      wait(50)
+      const { username } = await createReceiver({ fediversable: false })
+      const receiverActorUrl = makeActorUrl(username)
       await signedReq('get', endpoint, receiverActorUrl, emetterActorUrl, emetterUser.privateKey)
       .then(shouldNotBeCalled)
     } catch (err) {
@@ -53,8 +48,8 @@ describe('activitypub:actor', () => {
       const emetterUser = await createUserOnFediverse()
       emetterUser.publicKey = 'foo'
       const emetterActorUrl = await startServerWithEmetterUser(emetterUser)
-      const receiverActorUrl = await createReceiverActorUrl({ fediversable: false })
-      wait(50)
+      const { username } = await createReceiver({ fediversable: false })
+      const receiverActorUrl = makeActorUrl(username)
       await signedReq('get', endpoint, receiverActorUrl, emetterActorUrl, emetterUser.privateKey)
       .then(shouldNotBeCalled)
     } catch (err) {
@@ -72,7 +67,8 @@ describe('activitypub:actor', () => {
       const anotherUser = await createUserOnFediverse()
       emetterUser.privateKey = anotherUser.privateKey
       const emetterActorUrl = await startServerWithEmetterUser(emetterUser)
-      const receiverActorUrl = await createReceiverActorUrl({ fediversable: false })
+      const { username } = await createReceiver({ fediversable: false })
+      const receiverActorUrl = makeActorUrl(username)
       await signedReq('get', endpoint, receiverActorUrl, emetterActorUrl, emetterUser.privateKey)
       .then(shouldNotBeCalled)
     } catch (err) {
@@ -89,7 +85,7 @@ describe('activitypub:actor', () => {
       const emetterUser = await createUserOnFediverse()
       const emetterActorUrl = await startServerWithEmetterUser(emetterUser)
       const imaginaryReceiverUsername = createUsername()
-      const receiverActorUrl = `${host}${query(imaginaryReceiverUsername)}`
+      const receiverActorUrl = makeActorUrl(imaginaryReceiverUsername)
       await signedReq('get', endpoint, receiverActorUrl, emetterActorUrl, emetterUser.privateKey)
       .then(shouldNotBeCalled)
     } catch (err) {
@@ -105,7 +101,8 @@ describe('activitypub:actor', () => {
     try {
       const emetterUser = await createUserOnFediverse()
       const emetterActorUrl = await startServerWithEmetterUser(emetterUser)
-      const receiverActorUrl = await createReceiverActorUrl({ fediversable: false })
+      const { username } = await createReceiver({ fediversable: false })
+      const receiverActorUrl = makeActorUrl(username)
       await signedReq('get', endpoint, receiverActorUrl, emetterActorUrl, emetterUser.privateKey)
       .then(shouldNotBeCalled)
     } catch (err) {
@@ -120,7 +117,8 @@ describe('activitypub:actor', () => {
   it('should return a json ld file with a receiver actor url', async () => {
     const emetterUser = await createUserOnFediverse()
     const emetterActorUrl = await startServerWithEmetterUser(emetterUser)
-    const receiverActorUrl = await createReceiverActorUrl()
+    const { username } = await createReceiver()
+    const receiverActorUrl = makeActorUrl(username)
     const res = await signedReq('get', endpoint, receiverActorUrl, emetterActorUrl, emetterUser.privateKey)
     const body = JSON.parse(res.body)
     body['@context'].should.an.Array()
@@ -130,19 +128,3 @@ describe('activitypub:actor', () => {
     body.publicKey.owner.should.equal(receiverActorUrl)
   })
 })
-
-const createReceiverActorUrl = async (customData = {}) => {
-  const receiverUsername = createUsername()
-  const userAttributes = _.extend({
-    username: receiverUsername,
-    fediversable: true
-  }, customData)
-  await createUser(userAttributes)
-  const receiverActorUrl = `${host}${query(receiverUsername)}`
-  return receiverActorUrl
-}
-
-const startServerWithEmetterUser = async emetterUser => {
-  const { origin } = await startActivityPubServer(emetterUser)
-  return `${origin}${query(emetterUser.username)}`
-}
